refactor: refactor search component
