+++ src/components/SearchBar/SearchBar.tsx
@@ -1,28 +1,17 @@
-import React, { useEffect } from 'react';
-import { useAppDispatch, useAppSelector } from '@/app/hooks';
-import { setSearchText } from '@/features/laureates/laureatesSlice';
+import React from 'react';
+
 interface SearchBarProps {
   onSearch: (query: string) => void;
+  searchText: string;
 }
 
-const SearchBar: React.FC<SearchBarProps> = ({ onSearch }) => {
-  const dispatch = useAppDispatch();
-  const searchText = useAppSelector((state) => state.laureate.searchText);
-
+const SearchBar: React.FC<SearchBarProps> = ({ onSearch, searchText: defaultSearchText }) => {
+  const [searchText, setSearchText] = React.useState(defaultSearchText);
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
     onSearch(searchText);
     localStorage.setItem('searchQuery', searchText);
   };
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    dispatch(setSearchText(e.target.value.trim()));
-  };
-
-  useEffect(() => {
-    const savedQuery = localStorage.getItem('searchQuery') || '';
-    dispatch(setSearchText(savedQuery));
-  }, [dispatch]);
 
   return (
     <div className=" justify-start mb-4">
@@ -34,7 +23,7 @@ const SearchBar: React.FC<SearchBarProps> = ({ onSearch }) => {
           className="h-full w-full rounded border bg-transparent px-3 py-2.5 font-sans text-sm font-normal transition-all focus:border-pink-500 focus:outline-0"
           type="text"
           value={searchText}
-          onChange={handleChange}
+          onChange={(e) => setSearchText(e.target.value)}
           placeholder=" "
         />
         <button
