+++ src/pages/Laureates/Laureates.tsx
@@ -4,9 +4,13 @@ import { useSearchLaureatesQuery } from '@/api/laureates';
 import Pagination from '@/components/Pagination/Pagination';
 import LaureatesList from './LaureatesList';
 import Loading from '@/components/Loading';
-import {DEFAULT_LIMIT, DEFAULT_PAGE} from "@/config";
+import { DEFAULT_LIMIT, DEFAULT_PAGE } from '@/config';
+import { useAppDispatch, useAppSelector } from '@/app/hooks';
+import { setSearchText } from '@/features/laureates/laureatesSlice';
+import { useEffect } from 'react';
 
 export default function Laureates() {
+  const dispatch = useAppDispatch();
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
   const name = searchParams.get('search') ?? '';
@@ -16,6 +20,12 @@ export default function Laureates() {
   const req = { name, page, limit };
   const res = useSearchLaureatesQuery(req);
   const { data, error, isFetching, isSuccess, isError } = res;
+  const searchText = useAppSelector((state) => state.laureate.searchText);
+
+  useEffect(() => {
+    const savedQuery = localStorage.getItem('searchQuery') ?? '';
+    dispatch(setSearchText(savedQuery));
+  }, [dispatch]);
 
   const onSearch = (query: string) => {
     searchParams.set('search', query);
@@ -38,20 +48,19 @@ export default function Laureates() {
     });
   };
 
+  const createPageUrl = (searchParams: URLSearchParams, page: number) => {
+    const params = new URLSearchParams(searchParams);
+    params.set('page', page.toString());
+    return params.toString();
+  };
   let content;
 
   if (isFetching) {
     content = <Loading />;
   } else if (isSuccess) {
     const { total, laureates } = data;
-    const left = new URLSearchParams({
-      ...searchParams,
-      page: (page - 1).toString(),
-    }).toString();
-    const right = new URLSearchParams({
-      ...searchParams,
-      page: (page + 1).toString(),
-    }).toString();
+    const left = createPageUrl(searchParams, page - 1);
+    const right = createPageUrl(searchParams, page + 1);
     content = (
       <>
         <Pagination
@@ -75,7 +84,7 @@ export default function Laureates() {
   return (
     <div className="flex flex-row justify-start p-4">
       <div className="flex flex-col flex-grow mr-4">
-        <SearchBar onSearch={onSearch} />
+        <SearchBar onSearch={onSearch} searchText={searchText} />
         {content}
       </div>
       <div className="h-screen flex flex-col justify-center">
